make: Make XcodeSettings per-target instead of per-configuration, part 1.
This makes it possible to move all the bundle path logic into XcodeSettings,
which is required if xcode_config_file ever needs to be supported. It will
also make it easier to eventually write a ninja generator, and it's
required if GetLdFlags() needs to read the results of GetXcodeEnv() (which
will be moved into XcodeSettings, too).

BUG=http://crbug.com/90043, http://crbug.com/90199
Review URL: http://codereview.chromium.org/7493049

git-svn-id: https://gyp.googlecode.com/svn/trunk@978 78cadc50-ecff-11dd-a971-7dbc132099af