mac; Use libtool -static instead of ar crs to create static libraries.
Speed-up for libbrowser.a:
On 10.6 with a normal disk: from 27s to 10s
On 10.7 with an SSD: from 30s to 9s

Touching a single UI file and rebuilding chrome:
On 10.6 with a normal disk: from 120s to 100s
On 10.7 with an SSD: From 90s to 50s

ARFLAGS is still defined in the Makefile even on mac, but it doesn't end
up being used. Didn't seem worth the effort.
Review URL: http://codereview.chromium.org/7569016

git-svn-id: https://gyp.googlecode.com/svn/trunk@985 78cadc50-ecff-11dd-a971-7dbc132099af