ninja/mac: Support postbuilds.
Do this the same way the make build does it, but just tacking all
postbuilds on to the last command ran for a gyp target.

We considered several other options. This one has the advantages
to be simple and similar to what the make build does. If ninja
grows support for running multiple commands for an output, this
approach can be changed to take advantage of that fairly easily.

BUG=chromium:96894
Review URL: https://chromiumcodereview.appspot.com/9139083

git-svn-id: https://gyp.googlecode.com/svn/trunk@1180 78cadc50-ecff-11dd-a971-7dbc132099af