Add msvs generator flag to control _excluded_files inclusion.
Add a generator flag "msvs_list_excluded_files"
to not list _excluded_files in the generated project. They seem especially
annoying in VS2010 as they seem to get searched many times so Find in Files is
very noisy with duplicate results.

Default behaviour is still to include them, so there's only a change in
behaviour when that flag is set explicitly to 0. Similar flag for xcode is
"xcode_list_excluded_files".

Adds test for both 2008 and 2010.
Review URL: https://chromiumcodereview.appspot.com/9325038

git-svn-id: https://gyp.googlecode.com/svn/trunk@1187 78cadc50-ecff-11dd-a971-7dbc132099af